Fix an optimize the solver
Turns out one of the major optimizations is not usable with the bumpers.
Thankfully the heuristic for prioritizing which states to expand first
seems to be sufficient.
Changed the solver thread to fire a timeout, since I believe that it
needs to end the 'message' handling in a reasonable amount of time,
otherwise the parent process crashes or something. Also added some
explicit exception handling just in case that happens again so it
doesn't tank the entire game.